workbench-services/src/test/java/com/avalon/workbench/service/mock/concurrentReport/test/ConcurrentReportServiceImplTest.java
package com.avalon.workbench.service.mock.concurrentReport.test;

import java.util.ArrayList;
import java.util.List;

import org.apache.log4j.Logger;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.mockito.runners.MockitoJUnitRunner;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;

import com.avalon.workbench.beans.concurrntPrograms.ConcurrentPrograms;
import com.avalon.workbench.repository.concurrentReport.concurrentReportRepository;
import com.avalon.workbench.repository.concurrentReport.concurrentReportRepositoryImpl;
import com.avalon.workbench.services.concurrentReport.ConcurrentReportServiceImpl;
import com.avalon.workbench.services.test.AbstractWorkbenchServiceTest;

//@RunWith(MockitoJUnitRunner.class)
public class ConcurrentReportServiceImplTest extends
		AbstractWorkbenchServiceTest {
	protected static final Logger LOG_R = Logger
			.getLogger(ConcurrentReportServiceImplTest.class);
	
	@Autowired()
	@Qualifier(value = "ConcurrentReportServiceImpl")
	ConcurrentReportServiceImpl service;
	
	/*@InjectMocks
	private ConcurrentReportServiceImpl service;

	@Mock
	private concurrentReportRepositoryImpl concurrentReportRepositoryImpl;

	@Before
	public void init() {
		MockitoAnnotations.initMocks(this);
	}*/

	@Test
	public void test_runConcurrentProgram_Success() throws Exception {
		service.runConcurrentProgram("Purchasing, Vision Operations (USA)", "OPERATIONS", "EMPLOYEE_INFORMATION");
		Assert.assertTrue(true);
	}
	@Test
	public void test_getConcurrentReport_Success() throws Exception {
		service.getConcurrentReport();
		Assert.assertTrue(true);
	}

}
